Some more refactoring

+++ PingPayments.PaymentsApi.Tests/V1/BaseEndpointsTests.cs
@@ -22,7 +22,7 @@ namespace PingPayments.PaymentsApi.Tests.V1
         protected static void AssertHttpOK<T>(ApiResponseBase<T> response) where T : EmptySuccesfulResponseBody
         {
             Assert.NotNull(response);
-            Assert.Equal(200, response.StatusCode);
+            Assert.Equal(200, (int)response.StatusCode);
             Assert.False(response.IsFailure);
             Assert.True(response.IsSuccessful);
             Assert.NotNull(response.Body);
@@ -33,7 +33,7 @@ namespace PingPayments.PaymentsApi.Tests.V1
         protected static void AssertHttpNoContent<T>(ApiResponseBase<T> response) where T : EmptySuccesfulResponseBody
         {
             Assert.NotNull(response);
-            Assert.Equal(204, response.StatusCode);
+            Assert.Equal(204, (int)response.StatusCode);
             Assert.False(response.IsFailure);
             Assert.True(response.IsSuccessful);
             Assert.Null(response?.Body?.SuccesfulResponseBody);
@@ -43,7 +43,7 @@ namespace PingPayments.PaymentsApi.Tests.V1
         protected static void AssertHttpNotFound<T>(ApiResponseBase<T> response) where T : EmptySuccesfulResponseBody
         {
             Assert.NotNull(response);
-            Assert.Equal(404, response.StatusCode);
+            Assert.Equal(404, (int)response.StatusCode);
             Assert.True(response.IsFailure);
             Assert.False(response.IsSuccessful);
             Assert.Null(response?.Body?.SuccesfulResponseBody);
@@ -53,7 +53,7 @@ namespace PingPayments.PaymentsApi.Tests.V1
         protected static void AssertHttpUnprocessableEntity<T>(ApiResponseBase<T> response) where T : EmptySuccesfulResponseBody
         {
             Assert.NotNull(response);
-            Assert.Equal(422, response.StatusCode);
+            Assert.Equal(422, (int)response.StatusCode);
             Assert.True(response.IsFailure);
             Assert.False(response.IsSuccessful);
             Assert.Null(response?.Body?.SuccesfulResponseBody);
